backend/requirements.txt
alembic==1.16.2
fastapi[standard]>=0.112.0,<0.113.0
SQLAlchemy==2.0.41
pydantic==2.11.7
python-dotenv==1.1.1
pydantic-settings==2.10.1
pyjwt<3.0.0,>=2.8.0
python-jose==3.5.0
tenacity==9.1.2
passlib==1.7.4
psycopg2-binary==2.9.11
pytest
factory-boy==3.3.3
